@@ -9,7 +9,7 @@ import android.annotation.TargetApi;
 import android.app.IntentService;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 import com.jbdev.datamonitoring.utils.Logger;
 import java.util.Timer;
